feat(examples): common module for getting the org ID
