+++ utils-module-js/index.ts
@@ -1,0 +1,3 @@
+var nileUtils = require('./lib/nile');
+
+exports.nileUtils = nileUtils;
